pomelox_qwen_ai/requirements.txt
dashscope==1.20.4
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.1.1

# RAG 相关依赖
llama-index-core==0.10.67
llama-index-embeddings-dashscope==0.1.4
llama-index-readers-file==0.1.33
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
docx2txt==0.8
pydantic>=2.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0  # Excel导入支持
openpyxl>=3.1.0  # Excel读写

# 数据库驱动 (生产环境需要时安装)
pymysql>=1.1.0  # MySQL数据库驱动
sshtunnel>=0.4.0  # SSH隧道支持
paramiko<3.0  # SSH协议实现 (必须<3.0以兼容sshtunnel)
# psycopg2-binary>=2.9.0  # PostgreSQL (暂未使用)